zendesk/requirements.txt
llama-index
pinecone-client
python-dotenv
zenpy